feat: add Swagger API docs for all services with gateway aggregation; fix tenant lat/lng BigDecimal type mapping

+++ trip-service/src/main/java/com/travel/trip/dto/BrandingRequest.java
@@ -1,5 +1,7 @@
 package com.travel.trip.dto;
 
+import java.math.BigDecimal;
+
 public class BrandingRequest {
 
     private String logoUrl;
@@ -11,8 +13,8 @@ public class BrandingRequest {
     private String accentColor;
     private String tagline;
     private Integer orbIntensity;
-    private Double latitude;
-    private Double longitude;
+    private BigDecimal latitude;
+    private BigDecimal longitude;
     private String phone;
     private String address;
 
@@ -34,10 +36,10 @@ public class BrandingRequest {
     public void setTagline(String tagline) { this.tagline = tagline; }
     public Integer getOrbIntensity() { return orbIntensity; }
     public void setOrbIntensity(Integer orbIntensity) { this.orbIntensity = orbIntensity; }
-    public Double getLatitude() { return latitude; }
-    public void setLatitude(Double latitude) { this.latitude = latitude; }
-    public Double getLongitude() { return longitude; }
-    public void setLongitude(Double longitude) { this.longitude = longitude; }
+    public BigDecimal getLatitude() { return latitude; }
+    public void setLatitude(BigDecimal latitude) { this.latitude = latitude; }
+    public BigDecimal getLongitude() { return longitude; }
+    public void setLongitude(BigDecimal longitude) { this.longitude = longitude; }
     public String getPhone() { return phone; }
     public void setPhone(String phone) { this.phone = phone; }
     public String getAddress() { return address; }
